fix: Align ManifestInfo file count types with Iceberg v4 (#198)
* fix: change ManifestInfo file count types from Long to Integer

Iceberg v4 ManifestInfo defines added_files_count (504),
existing_files_count (505), and deleted_files_count (506) as
IntegerType (i32). The kernel was writing these as i64 (Long),
causing Parquet type mismatch errors when Iceberg reads manifests
since Parquet does not support Long to Integer demotion.

* Extract helper

* Fix

+++ kernel/src/content_tree/reader.rs
@@ -168,11 +168,11 @@ fn visit_metadata_entry_at<'a>(
     // content_stats has no fields, so no getters
 
     // Extract manifest_info fields (9 fields: 17-25, including dv and dv_cardinality)
-    let ms_added_files_count: Option<i64> =
+    let ms_added_files_count: Option<i32> =
         getters[17].get_opt(row_index, "manifest_info.added_files_count")?;
-    let ms_existing_files_count: Option<i64> =
+    let ms_existing_files_count: Option<i32> =
         getters[18].get_opt(row_index, "manifest_info.existing_files_count")?;
-    let ms_deletes_files_count: Option<i64> =
+    let ms_deletes_files_count: Option<i32> =
         getters[19].get_opt(row_index, "manifest_info.deletes_files_count")?;
     let ms_added_rows_count: Option<i64> =
         getters[20].get_opt(row_index, "manifest_info.added_rows_count")?;
